refactor: 🔨FilterUseCase로 뷰모델의 체크 로직 분리(#120)
